Move adjustedResults within search into mixin for shared use across categoryPreview & searchResults
